requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
faiss-cpu
openai
python-dotenv
numpy
